refactor: extract log buffer from drivers
