+++ lib/drivers/direct/container.js
@@ -4,12 +4,12 @@ var App = require('strong-runner').App;
 var Debug = require('debug');
 var _ = require('lodash');
 var assert = require('assert');
-var bl = require('bl');
 var c2s = require('strong-runner').Runnable.toString;
 var cicada = require('strong-fork-cicada');
 var cicadaCommit = require('strong-fork-cicada/lib/commit');
 var fs = require('fs');
 var localDeploy = require('../common/local-deploy');
+var logBuffer = require('../common/log-buffer');
 var mandatory = require('./../../util').mandatory;
 var packReceiver = require('../common/pack-receiver');
 var path = require('path');
@@ -42,12 +42,8 @@ function Container(options) {
   this.debug = Debug('strong-pm:container:' + this.instanceId);
 
   //this.restartCount = 0; // XXX(sam) move here from strong-runner.App?
-  this._appLog = bl();
-  this._logGcTimer = setInterval(
-    this._gcLogs.bind(this),
-    exports.LOG_GC_INTERVAL_MS
-  );
-  this._logGcTimer.unref();
+  this._appLog = logBuffer();
+  this._appLog.enableGC();
 
   this.stdout.pipe(this._appLog, {end: false});
   this.stderr.pipe(this._appLog, {end: false});
@@ -304,21 +300,11 @@ Container.prototype.updateInstanceEnv = function(env, callback) {
   self.emit('prepared', commit);
 };
 
-// This is a soft limit that is only checked/enforced once per LOG_GC interval.
-exports.MAX_LOG_RETENTION_BYTES = 1 * 1024 * 1024;
-exports.LOG_GC_INTERVAL_MS = 30 * 1000;
-
-Container.prototype._gcLogs = function() {
-  var overflow = this._appLog.length - exports.MAX_LOG_RETENTION_BYTES;
-  if (overflow > 0) {
-    this._appLog.consume(overflow);
+Container.prototype.dumpLogs = function() {
+  if (this.current) {
+    return this._appLog.dump();
+  } else {
+    return null;
   }
-};
 
-Container.prototype.readableLogSnapshot = function() {
-  return this._appLog.duplicate();
-};
-
-Container.prototype.flushLogs = function() {
-  this._appLog.consume(this._appLog.length);
 };
