+++ lib/drivers/direct/index.js
@@ -110,28 +110,11 @@ Driver.prototype.removeInstance = function(instanceId, callback) {
 
 Driver.prototype.dumpInstanceLog = function(instanceId) {
   var container = this._containerById(instanceId);
-  var log = container.readableLogSnapshot().toString();
-  var logLength = log === null ? undefined : log.length;
+  debug('dumpInstanceLog(%j) size: %j current? %j child? %j',
+        instanceId, current, child);
   var current = !!container.current;
   var child = current ? !!container.current.child : undefined;
-
-  container.flushLogs();
-
-  debug('dumpInstanceLog(%j) size: %j current? %j child? %j',
-        instanceId, logLength, current, child);
-
-  // Distinguish between no log data because there is no data, and no log
-  // data because there is no running child by setting log data to empty
-  // string.
-  /* eslint eqeqeq:0 */
-  if (log == '' && !child) {
-    log = null;
-  }
-  if (log == null && child) {
-    log = '';
-  }
-
-  return log;
+  return container.dumpLogs();
 };
 
 Driver.prototype.updateInstanceEnv = function(instanceId, env, callback) {
