+++ lib/drivers/docker/index.js
@@ -2,7 +2,6 @@
 
 var _ = require('lodash');
 var async = require('async');
-var bl = require('bl');
 var debug = require('debug')('strong-pm:docker:driver');
 var Docker = require('dockerode');
 var EventEmitter = require('events').EventEmitter;
@@ -10,6 +9,7 @@ var fmt = require('util').format;
 var Image = require('./image');
 var inherits = require('util').inherits;
 var isError = require('util').isError;
+var logBuffer = require('../common/log-buffer');
 var mandatory = require('./../../util').mandatory;
 
 module.exports = exports = DockerDriver;
@@ -195,13 +195,12 @@ DockerDriver.prototype.stopInstance = function(id, style, cb) {
 DockerDriver.prototype.dumpInstanceLog = function(id) {
   debug('DockerDriver.dumpInstanceLog(%j)', id);
   var instance = this._instance(id);
-  if (!instance.current) {
+  if (instance.current) {
+    return instance.log.dump();
+  } else {
     debug('no instance to dump logs for: %j', id);
     return null;
   }
-  var logDump = instance.log.duplicate().toString();
-  instance.log.consume(logDump.length);
-  return logDump;
 };
 
 DockerDriver.prototype.updateInstanceEnv = function(id, env, cb) {
@@ -337,42 +336,10 @@ DockerDriver.prototype._instance =
 DockerDriver.prototype._containerById = function(id) {
   this._instances[id] = this._instances[id] || {
     startOpts: _.clone(this.defaultStartOptions),
-    log: makeLogBuffer(),
+    log: logBuffer(),
   };
   return this._instances[id];
 };
-
-// TODO: extract this log buffer to drivers/common
-// This is a soft limit that is only checked/enforced once per LOG_GC interval.
-exports.MAX_LOG_RETENTION_BYTES = 1 * 1024 * 1024;
-exports.LOG_GC_INTERVAL_MS = 30 * 1000;
-
-function makeLogBuffer() {
-  var logger = bl();
-  logger.enableGC = function() {
-    logger._logGcTimer = logger._logGcTimer || makeGcTimer();
-  };
-  logger.on('end', function() {
-    if (logger._logGcTimer) {
-      clearInterval(logger.logGcTimer);
-      logger._logGcTimer = null;
-    }
-  });
-  return logger;
-
-  function gcLogger() {
-    var overflow = logger.length - exports.MAX_LOG_RETENTION_BYTES;
-    if (overflow > 0) {
-      logger.consume(overflow);
-    }
-  }
-
-  function makeGcTimer() {
-    var timer = setInterval(gcLogger, exports.LOG_GC_INTERVAL_MS);
-    timer.unref();
-    return timer;
-  }
-}
 
 DockerDriver.prototype.getName = function() {
   return 'Docker';
